+++ src/features/ui/post/PostEditDialog.tsx
@@ -1,5 +1,5 @@
 import { Dialog, DialogContent, DialogHeader, DialogTitle, Input, Textarea, Button } from "../../../shared/ui"
-import { usePost } from "../../../features/model/post/usePost"
+import { usePost } from "../../model/post/usePost"
 
 export const PostEditDialog = () => {
   const { showEditDialog, setShowEditDialog, handleUpdatePost, setSelectedPost, selectedPost } = usePost()
