+++ src/pages/PostsManagerPage.tsx
@@ -4,13 +4,13 @@ import { URLParams } from "../shared/types"
 import { useTag } from "../features/model/tag/useTag"
 import { usePost } from "../features/model/post/usePost"
 import { useURLParams } from "../features/model/url/useURLParams"
-import { PostSearchInput } from "../widgets/ui/post/PostSearchInput"
-import { PostTagSelect } from "../widgets/ui/post/PostTagSelect"
-import { PostSortSelect } from "../widgets/ui/post/PostSortSelect"
+import { PostSearchInput } from "../features/ui/post/PostSearchInput"
+import { PostTagSelect } from "../features/ui/post/PostTagSelect"
+import { PostSortSelect } from "../features/ui/post/PostSortSelect"
 import { PostTable } from "../widgets/ui/post/PostTable"
 import { PostPagination } from "../widgets/ui/post/PostPagination"
-import { PostAddDialog } from "../widgets/ui/post/PostAddDialog"
-import { PostEditDialog } from "../widgets/ui/post/PostEditDialog"
+import { PostAddDialog } from "../features/ui/post/PostAddDialog"
+import { PostEditDialog } from "../features/ui/post/PostEditDialog"
 import { PostDetailDialog } from "../widgets/ui/post/PostDetailDialog"
 import { CommentAddDialog } from "../features/ui/comment/CommentAddDialog"
 import { CommentEditDialog } from "../features/ui/comment/CommentEditDialog"
