+++ src/widgets/ui/post/PostFilters.tsx
@@ -1,6 +1,6 @@
-import { PostSearchInput } from "./PostSearchInput"
-import { PostTagSelect } from "./PostTagSelect"
-import { PostSortSelect } from "./PostSortSelect"
+import { PostSearchInput } from "../../../features/ui/post/PostSearchInput"
+import { PostTagSelect } from "../../../features/ui/post/PostTagSelect"
+import { PostSortSelect } from "../../../features/ui/post/PostSortSelect"
 import { useURLParams } from "../../../features/model/url/useURLParams"
 import { usePost } from "../../../features/model/post/usePost"
 import { useTag } from "../../../features/model/tag/useTag"
